pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rails-migration-assistant"
version = "1.0.0"
description = "AI-powered Rails upgrade assistance with dual LLM support for secure modernization"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Mohammed Hassan", email = "your.email@example.com"}
]
maintainers = [
    {name = "Mohammed Hassan", email = "your.email@example.com"}
]
keywords = [
    "rails", "ruby", "upgrade", "migration", "ai", "llm", 
    "modernization", "automation", "assistant", "gemini", "local-llm"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Ruby",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "gitpython>=3.1.0",
    "python-dotenv>=0.19.0",
    "google-generativeai>=0.3.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "bitsandbytes>=0.41.0",
    "accelerate>=0.20.0",
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "lxml>=4.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
]
gui = [
    "tkinter-page>=7.0.0",
]
performance = [
    "auto-gptq>=0.4.0",
    "optimum>=1.12.0",
    "flash-attn>=2.0.0",
    "xformers>=0.0.20",
]

[project.scripts]
rails-migrate = "rails_upgrade_suggestions:main"
rails-gui = "rails_upgrade_gui:main"

[project.urls]
Homepage = "https://github.com/your-username/rails-migration-assistant"
Documentation = "https://github.com/your-username/rails-migration-assistant#readme"
Repository = "https://github.com/your-username/rails-migration-assistant"
"Bug Tracker" = "https://github.com/your-username/rails-migration-assistant/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.json", "*.jsonl", "*.index", "*.patch"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
